adds init e2e tests for homepage
